+++ lib/src/platform/messages.g.dart
@@ -17,6 +17,17 @@ PlatformException _createConnectionError(String channelName) {
   );
 }
 
+List<Object?> wrapResponse(
+    {Object? result, PlatformException? error, bool empty = false}) {
+  if (empty) {
+    return <Object?>[];
+  }
+  if (error == null) {
+    return <Object?>[result];
+  }
+  return <Object?>[error.code, error.message, error.details];
+}
+
 class _PigeonCodec extends StandardMessageCodec {
   const _PigeonCodec();
   @override
@@ -52,4 +63,49 @@ class ClixHostApi {
   static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
 
   final String pigeonVar_messageChannelSuffix;
+}
+
+abstract class ClixFlutterApi {
+  static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
+
+  void onNotificationTapped(Map<String?, Object?> userInfo);
+
+  static void setUp(
+    ClixFlutterApi? api, {
+    BinaryMessenger? binaryMessenger,
+    String messageChannelSuffix = '',
+  }) {
+    messageChannelSuffix =
+        messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
+    {
+      final BasicMessageChannel<
+          Object?> pigeonVar_channel = BasicMessageChannel<
+              Object?>(
+          'dev.flutter.pigeon.clix_flutter.ClixFlutterApi.onNotificationTapped$messageChannelSuffix',
+          pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
+      if (api == null) {
+        pigeonVar_channel.setMessageHandler(null);
+      } else {
+        pigeonVar_channel.setMessageHandler((Object? message) async {
+          assert(message != null,
+              'Argument for dev.flutter.pigeon.clix_flutter.ClixFlutterApi.onNotificationTapped was null.');
+          final List<Object?> args = (message as List<Object?>?)!;
+          final Map<String?, Object?>? arg_userInfo =
+              (args[0] as Map<Object?, Object?>?)?.cast<String?, Object?>();
+          assert(arg_userInfo != null,
+              'Argument for dev.flutter.pigeon.clix_flutter.ClixFlutterApi.onNotificationTapped was null, expected non-null Map<String?, Object?>.');
+          try {
+            api.onNotificationTapped(arg_userInfo!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          } catch (e) {
+            return wrapResponse(
+                error: PlatformException(code: 'error', message: e.toString()));
+          }
+        });
+      }
+    }
+  }
 }